requirements.txt
flask
pandas
numpy
scikit-learn
matplotlib
psycopg2
psycopg2-binary
gunicorn
python-dotenv
